+++ src/AppInstallerCommonCore/Public/winget/Manifest.h
@@ -7,6 +7,7 @@
 #include <winget/ManifestInstaller.h>
 #include <winget/ManifestLocalization.h>
 
+#include <utility>
 #include <vector>
 
 namespace AppInstaller::Manifest
@@ -63,6 +64,9 @@ namespace AppInstaller::Manifest
 
         // Get publishers across localizations and installers, Case folded.
         std::vector<string_t> GetPublishers() const;
+
+        // Gets name/publisher pairs with default fallbacks, preserving case.
+        std::vector<std::pair<string_t, string_t>> GetNameAndPublisherPairs() const;
 
         // If not empty, the SHA256 hash of the manifest stream itself.
         Utility::SHA256::HashBuffer StreamSha256;